requirements.txt
# ==============================================
# DEPENDENCIAS PRINCIPALES
# ==============================================

# API y HTTP
requests>=2.31.0
httpx>=0.24.1
aiohttp>=3.8.5

# Blockchain y Web3
web3>=6.0.0
eth-account>=0.9.0

# MODO EXECUTE - Trading automation
py-clob-client>=0.12.0

# Base de datos
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Configuración
python-dotenv>=1.0.0

# Utilidades
python-dateutil>=2.8.2
pytz>=2023.3

# ==============================================
# V2.0 - ML GAP PREDICTOR
# ==============================================

# Machine Learning
tensorflow>=2.15.0
keras>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Sentiment Analysis
transformers>=4.35.0
torch>=2.1.0
tweepy>=4.14.0
praw>=7.7.0
vaderSentiment>=3.3.2

# Data Processing
nltk>=3.8.1
spacy>=3.7.0
textblob>=0.17.1

# Notifications
python-telegram-bot>=20.4

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ==============================================
# OPCIONAL (desarrollo y pruebas)
# ==============================================

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Linting y formateo
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Type hints
types-requests>=2.31.0
types-python-dateutil>=2.8.19

# ==============================================
# PRODUCTION (opcional)
# ==============================================

# Monitoring
# sentry-sdk>=1.30.0  # Uncomment para error tracking
